Require Tesseract 3.2 or greater, generate HOCR rather than RAW_HOCR/ENCODED_OCR
Versions of Tesseract prior to 3.2 generated HTML 4.01 and for some files it
would generate invalid HTML 4.01 when creating HOCR. Setting the restriction
that this module use Tesseract 3.2 or greater resolves these problems.

Tesseract 3.2 generates valid XHTML so it need not be encoded, also there is
more relevant coordinate information generated by it.

This committed changes the DC_COMPOSITE_MODEL for pages and updates the admin
form to state the required version and if it is installed.